backend/requirements.txt
fastapi
uvicorn
pydantic-settings
pydantic
python-dotenv
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
bcrypt